Fix data type

+++ Hazel/src/Platform/Windows/WindowsWindow.cpp
@@ -119,13 +119,13 @@ namespace Hazel {
 		glfwSetScrollCallback(m_Window, [](GLFWwindow* window, double xOffset, double yOffset)
 			{
 				WindowData& data = *(WindowData*)glfwGetWindowUserPointer(window);
-				MouseScrolledEvent event(xOffset,yOffset);
+				MouseScrolledEvent event((float)xOffset,(float)yOffset);
 				data.EventCallback(event);
 			});
 		glfwSetCursorPosCallback(m_Window, [](GLFWwindow* window, double xPos, double yPos)
 			{
 				WindowData& data = *(WindowData*)glfwGetWindowUserPointer(window);
-				MouseMovedEvent event(xPos, yPos);
+				MouseMovedEvent event((float)xPos, (float)yPos);
 				data.EventCallback(event);
 			});
 		glfwSetCharCallback(m_Window, [](GLFWwindow* window, unsigned int keycode)
